+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### Changed
+- add fallback for MONGODB and REDIS connection strings, fix logging
 - create MONGODB connection string from environment variables instead of full MONGO_URI variable
 - create REDIS connection string from environment variables instead of full REDIS_URL variable