+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # Changelog
 
 ### Changed
+
+## [1.2.0]
+
+### Changed
 - add fallback for MONGODB and REDIS connection strings, fix logging
 - create MONGODB connection string from environment variables instead of full MONGO_URI variable
 - create REDIS connection string from environment variables instead of full REDIS_URL variable
+- add authorization mechanism
 
-## 1.1.0
+## [1.1.0]
 
 ### Changed
 - add error handling for apply changes action